CODE RUB:  Code cleanup
