+++ STX.REST.RESTFulSense.Clients/Brokers/Errors/ErrorBroker.StatusDetails.cs
@@ -3,9 +3,9 @@
 // ----------------------------------------------------------------------------------
 
 using System.Collections.Generic;
-using STX.REST.RESTFulSense.Clients.Models.Brokers.StatusDetails;
+using STX.REST.RESTFulSense.Clients.Models.Brokers.Errors;
 
-namespace STX.REST.RESTFulSense.Clients.Brokers.ErrorBrokers
+namespace STX.REST.RESTFulSense.Clients.Brokers.Errors
 {
     internal partial class ErrorBroker : IErrorBroker
     {
