+++ STX.REST.RESTFulSense.Clients/Models/Brokers/Errors/StatusDetail.cs
@@ -2,7 +2,7 @@
 // Copyright (c) The Standard Organization: A coalition of the Good-Hearted Engineers
 // ----------------------------------------------------------------------------------
 
-namespace STX.REST.RESTFulSense.Clients.Models.Brokers.StatusDetails
+namespace STX.REST.RESTFulSense.Clients.Models.Brokers.Errors
 {
     internal class StatusDetail
     {
